# swagger.json
{"openapi":"3.0.0","info":{"title":"API","version":"1.0.0","description":"API documentation for endpoints"},"components":{"schemas":{"LoginDTO":{"properties":{"email":{"format":"email","type":"string"},"password":{"minLength":8,"maxLength":20,"type":"string"}},"type":"object","required":["email","password"]},"LockCouponDTO":{"properties":{"userId":{"minLength":1,"type":"string"}},"type":"object","required":["userId"]},"AssignCouponDTO":{"properties":{"userId":{"minLength":1,"type":"string"}},"type":"object","required":["userId"]},"AssignRandomCouponDTO":{"properties":{"userId":{"minLength":1,"type":"string"},"couponBookId":{"minLength":1,"type":"string"}},"type":"object","required":["userId","couponBookId"]},"CreateCouponBookDTO":{"properties":{"name":{"type":"string"},"description":{"type":"string"},"startDate":{"type":"string"},"endDate":{"type":"string"},"isActive":{"type":"boolean"},"maxRedemptionsPerUser":{"type":"number"},"maxCodesPerUsers":{"type":"number"},"allowMultipleRedemptions":{"type":"boolean"},"codePattern":{"type":"string"}},"type":"object","required":["name","startDate","endDate"]},"GenerateCodesDTO":{"properties":{"quantity":{"minLength":1,"type":"number","minimum":1},"couponBookId":{"minLength":1,"type":"string"}},"type":"object","required":["quantity","couponBookId"]},"UploadCouponsDTO":{"properties":{"codes":{"minLength":1,"type":"array","items":{},"minItems":1},"couponBookId":{"minLength":1,"type":"string"}},"type":"object","required":["codes","couponBookId"]},"CreateUserDTO":{"properties":{"firstName":{"minLength":3,"maxLength":20,"type":"string"},"lastName":{"minLength":3,"maxLength":20,"type":"string"},"email":{"format":"email","type":"string"},"password":{"minLength":8,"maxLength":20,"type":"string"}},"type":"object","required":["firstName","lastName","email","password"]}},"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"security":[{"BearerAuth":[]}],"externalDocs":{"description":"swagger.json","url":"/swagger.json"},"servers":[{"url":"http://localhost:3000"}],"paths":{"/auth/token":{"post":{"summary":"Login user and get JWT token","tags":["Auth"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"200":{"description":"Successful login"},"401":{"description":"Invalid credentials"},"404":{"description":"User not found"}}}},"/coupons":{"post":{"summary":"Create a Coupon Book","tags":["Coupon Book"],"security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCouponBookDTO"}}}},"responses":{"201":{"description":"Coupon Book created successfully"},"400":{"description":"Bad request"}}}},"/coupons/random-codes":{"post":{"summary":"Generate coupons for book","tags":["Coupon Book"],"security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateCodesDTO"}}}},"responses":{"201":{"description":"Coupon Book created successfully"},"400":{"description":"Bad request"}}}},"/coupons/codes":{"post":{"summary":"Upload coupons for book","tags":["Coupon Book"],"security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadCouponsDTO"}}}},"responses":{"201":{"description":"Coupons upload successfully"},"400":{"description":"Bad request"}}}},"/coupons/assign":{"post":{"summary":"Assign random coupon to user","tags":["Coupons"],"security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignRandomCouponDTO"}}}},"responses":{"201":{"description":"Coupon assigned successfully"},"400":{"description":"Bad request"}}}},"/coupons/assign/{code}":{"post":{"summary":"Assign coupon to user","tags":["Coupons"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"},"description":"Coupon code"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignCouponDTO"}}}},"responses":{"201":{"description":"Coupon assigned successfully"},"400":{"description":"Bad request"}}}},"/coupons/lock/{code}":{"post":{"summary":"Lock a coupon for redemption","tags":["Coupons"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"},"description":"Coupon code"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LockCouponDTO"}}}},"responses":{"201":{"description":"Coupon locked successfully"},"400":{"description":"Bad request"}}}},"/coupons/redeem/{code}":{"post":{"summary":"redeem a coupon","tags":["Coupons"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"},"description":"Coupon code"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignCouponDTO"}}}},"responses":{"201":{"description":"Coupon locked successfully"},"400":{"description":"Bad request"}}}},"/coupons/user/{id}":{"get":{"summary":"Return coupons assigned to user","tags":["Coupons"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"description":"User id"}],"responses":{"201":{"description":"Coupon locked successfully"},"400":{"description":"Bad request"}}}},"/users":{"post":{"summary":"Create a user","tags":["Users"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDTO"}}}},"responses":{"201":{"description":"User created successfully"},"400":{"description":"Bad request"}}}}},"tags":[{"name":"Auth","description":"Authentication endpoints"},{"name":"Coupon Book","description":"Coupon Book management endpoints"},{"name":"Coupons","description":"Coupons management endpoints"},{"name":"Users","description":"User management endpoints"}]}